update partition id dtype

+++ kahypar-resources/definitions.h
@@ -39,7 +39,7 @@ using HypernodeID = uint32_t;
 using HyperedgeID = uint32_t;
 using HypernodeWeight = int64_t;
 using HyperedgeWeight = int64_t;
-using PartitionID = int32_t;
+using PartitionID = int64_t;
 using Gain = HyperedgeWeight;
 
 // #########Graph-Definitions#############
